Renamed Hijack to Recovery, updated code base and changed narrative
