+++ gamedata/descriptions/levels/recovery.md
@@ -1,0 +1,4 @@
+
+A contract creator has built a very simple token factory contract. Anyone can create new tokens with ease. After deploying the first token contract, the creator sent `0.5` ether to obtain more tokens. They have since lost the contract address. 
+
+This level will be completed if you can recover (or remove) the `0.5` ether from the lost contract address.
